Removed Secret Manager code  (#6481)
