+++ javav2/example_code/secrets-manager/src/main/resources/config.properties
