--- javav2/example_code/secretsmanager/metadata.yaml
@@ -1,23 +1,0 @@
---- 
-files:
- - path: src/main/java/com/example/secrets/CreateSecret.java
-   services:
-     - secretsmanager
- - path: src/main/java/com/example/secrets/DeleteSecret.java
-   services:
-     - secretsmanager
- - path: src/main/java/com/example/secrets/DescribeSecret.java
-   services:
-     - secretsmanager
- - path: src/main/java/com/example/secrets/GetSecretValue.java
-   services:
-     - secretsmanager
- - path: src/main/java/com/example/secrets/ListSecrets.java
-   services:
-     - secretsmanager
- - path: src/main/java/com/example/secrets/UpdateSecret.java
-   services:
-     - secretsmanager
- - path: src/test/java/SecretManagerTest.java
-   services:
-     - secretsmanager
