--- javav2/example_code/secretsmanager/src/main/java/com/example/secrets/UpdateSecret.java
@@ -1,65 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.secrets;
-
-// snippet-start:[secretsmanager.java2.update_secret.main]
-// snippet-start:[secretsmanager.java2.update_secret.import]
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
-import software.amazon.awssdk.services.secretsmanager.model.SecretsManagerException;
-import software.amazon.awssdk.services.secretsmanager.model.UpdateSecretRequest;
-// snippet-end:[secretsmanager.java2.update_secret.import]
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class UpdateSecret {
-    public static void main(String[] args) {
-        final String usage = """
-
-                Usage:
-                    <secretName> <secretValue>
-
-                Where:
-                    secretName - The name of the secret (for example, tutorials/MyFirstSecret).\s
-                    secretValue - The secret value that is updated.\s
-                """;
-
-        if (args.length < 2) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String secretName = args[0];
-        String secretValue = args[1];
-        Region region = Region.US_EAST_1;
-        SecretsManagerClient secretsClient = SecretsManagerClient.builder()
-                .region(region)
-                .build();
-
-        updateMySecret(secretsClient, secretName, secretValue);
-        secretsClient.close();
-    }
-
-    public static void updateMySecret(SecretsManagerClient secretsClient, String secretName, String secretValue) {
-        try {
-            UpdateSecretRequest secretRequest = UpdateSecretRequest.builder()
-                    .secretId(secretName)
-                    .secretString(secretValue)
-                    .build();
-
-            secretsClient.updateSecret(secretRequest);
-
-        } catch (SecretsManagerException e) {
-            System.err.println(e.awsErrorDetails().errorMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[secretsmanager.java2.update_secret.main]
